requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
PyHamcrest==2.0.3
pyparsing==3.0.7
pytest==6.2.5
requests==2.27.1
toml==0.10.2
urllib3==1.26.8
